src/agentservice/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2
google-cloud-aiplatform>=1.43.0
google-genai>=0.8.0
vertexai>=1.43.0
python-multipart==0.0.6 
python-dotenv==1.0.0 